requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
confusable-homoglyphs==2.0.2
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
Django==1.11.7
django-annoying==0.10.3
django-autoslug==1.9.3
django-cities-light==3.3.0
django-comments-xtd==2.0.9
django-contrib-comments==1.8.0
django-crispy-forms==1.7.0
django-map-widgets==0.1.9
django-registration==2.3
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
docutils==0.14
geoip2==2.6.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.1.1
maxminddb==1.3.0
mysqlclient==1.3.12
olefile==0.44
parsel==1.2.0
Pillow==4.3.0
progressbar2==3.34.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.4.0
python-utils==2.2.0
pytz==2017.3
queuelib==1.4.2
requests==2.18.4
Scrapy==1.4.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
Unidecode==0.4.21
urllib3==1.22
w3lib==1.18.0
zope.interface==4.4.3
